% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\docType{class}
\name{AbstractFormula-class}
\alias{AbstractFormula-class}
\alias{AbstractFormula}
\title{Abstracto formula}
\description{
Abstracto formula
}
\section{Slots}{

\describe{
\item{\code{type}}{character.}

\item{\code{name}}{character.}

\item{\code{arguments}}{list.}

\item{\code{string}}{character.}
}}

