% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_dynwrap.R
\name{add_regulatory_network}
\alias{add_regulatory_network}
\title{Add a GRN to a dynwrap object}
\usage{
add_regulatory_network(
  dataset,
  regulatory_network,
  regulatory_network_sc = NULL,
  regulators = NULL,
  targets = NULL,
  ...
)
}
\arguments{
\item{dataset}{A dataset created by \code{\link[dynwrap:wrap_data]{wrap_data()}} or \code{\link[dynwrap:wrap_expression]{wrap_expression()}}}

\item{regulatory_network}{A data frame consisting of three columns: \code{"regulator"}, \code{"target"}, \code{"strength"}.}

\item{regulatory_network_sc}{A data frame consisting of four columns: \code{"cell_id"}, \code{"regulator"}, \code{"target"}, \code{"strength"}.}

\item{regulators}{The feature ids of the regulators.}

\item{targets}{The feature ids of the targets.}

\item{...}{Extra arguments to be saved in the model.}
}
\description{
Add a GRN to a dynwrap object
}
