% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_network_generation.R
\name{add_targets_realnet}
\alias{add_targets_realnet}
\title{Add targets based on a real network}
\usage{
add_targets_realnet(net, geneinfo, realnet_name = "regulatorycircuits",
  damping = 0.05, ntargets_sampler = function() {     sample(20:100, 1) },
  gene_name_generator = function(i) paste0("G", i))
}
\arguments{
\item{net}{Network dataframe}

\item{geneinfo}{Geneinfo dataframe}

\item{realnet_name}{Name of the real network}

\item{damping}{Damping factor for personalized pagerank}
}
\description{
Add targets based on a real network
}
