% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\docType{data}
\name{base_params}
\alias{base_params}
\alias{simple_params}
\title{Base params}
\format{An object of class \code{list} of length 5.}
\usage{
base_params

simple_params
}
\description{
Base params
}
\keyword{datasets}
