% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_experiment.R
\name{check_expression}
\alias{check_expression}
\title{Checks the expression for certain properties}
\usage{
check_expression(expression)
}
\arguments{
\item{expression}{Expression matrix}
}
\description{
Checks the expression for certain properties
}
