% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\docType{methods}
\name{+,AbstractFormula,AbstractFormula-method}
\alias{+,AbstractFormula,AbstractFormula-method}
\alias{-,AbstractFormula,AbstractFormula-method}
\alias{*,AbstractFormula,AbstractFormula-method}
\alias{/,AbstractFormula,AbstractFormula-method}
\alias{+,AbstractFormula,numeric-method}
\alias{-,AbstractFormula,numeric-method}
\alias{*,AbstractFormula,numeric-method}
\alias{/,AbstractFormula,numeric-method}
\alias{+,numeric,AbstractFormula-method}
\alias{-,numeric,AbstractFormula-method}
\alias{*,numeric,AbstractFormula-method}
\alias{/,numeric,AbstractFormula-method}
\title{Short hand notations for generating new formulae}
\usage{
\S4method{+}{AbstractFormula,AbstractFormula}(e1, e2)

\S4method{-}{AbstractFormula,AbstractFormula}(e1, e2)

\S4method{*}{AbstractFormula,AbstractFormula}(e1, e2)

\S4method{/}{AbstractFormula,AbstractFormula}(e1, e2)

\S4method{+}{AbstractFormula,numeric}(e1, e2)

\S4method{-}{AbstractFormula,numeric}(e1, e2)

\S4method{*}{AbstractFormula,numeric}(e1, e2)

\S4method{/}{AbstractFormula,numeric}(e1, e2)

\S4method{+}{numeric,AbstractFormula}(e1, e2)

\S4method{-}{numeric,AbstractFormula}(e1, e2)

\S4method{*}{numeric,AbstractFormula}(e1, e2)

\S4method{/}{numeric,AbstractFormula}(e1, e2)
}
\arguments{
\item{e1}{Argument 1}

\item{e2}{Argument 2}
}
\value{
A new formula
}
\description{
Short hand notations for generating new formulae
}
