% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_gs.R
\name{divide_simulation_precise}
\alias{divide_simulation_precise}
\title{This function looks to make the rough estimates of splits and convergences more accurate
A bifurcation will be moved forward, a convergence will be moved backward}
\usage{
divide_simulation_precise(simulation, states, rough)
}
\description{
This function looks to make the rough estimates of splits and convergences more accurate
A bifurcation will be moved forward, a convergence will be moved backward
}
