% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example_model.R
\docType{data}
\name{example_model}
\alias{example_model}
\title{A (very!) small toy dyngen model}
\format{
An object of class \code{list} (inherits from \code{dyngen::init}) of length 19.
}
\usage{
example_model
}
\description{
Used for showcasing examples of functions.
}
\keyword{datasets}
