% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_goldstandard.R
\name{extract_goldstandard}
\alias{extract_goldstandard}
\title{Extract the gold standard}
\usage{
extract_goldstandard(simulation, model, reference_length = 100,
  max_path_length = 10, smooth_window = 50, verbose = TRUE,
  preprocess = TRUE)
}
\arguments{
\item{simulation}{The simulation}

\item{model}{The model}

\item{reference_length}{The length in the reference of each population}

\item{max_path_length}{Maximal number of milestones per path}

\item{smooth_window}{The window (steps) to smooth over}

\item{verbose}{Whether to include progress information}

\item{preprocess}{Whether to preprocess}
}
\description{
Extract the gold standard
}
