% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_network_generation.R
\name{extract_induced_subgraph_from_tf}
\alias{extract_induced_subgraph_from_tf}
\title{Extract subnetwork
Given a network, extract an induced subgraph from from a central TF to (indirect) targets using Personalized PageRank}
\usage{
extract_induced_subgraph_from_tf(net, tf_of_interest = NULL, damping = 0.05,
  ngenesampler = function() {     sample(20:100, 1) })
}
\arguments{
\item{net}{Network as a data.frame or igraph object}

\item{damping}{Damping factor for personalized pagerank}

\item{ngenesampler}{Function for sampling the number of genes}
}
\description{
Extract subnetwork
Given a network, extract an induced subgraph from from a central TF to (indirect) targets using Personalized PageRank
}
