% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_goldstandard.R
\name{extract_references}
\alias{extract_references}
\title{Extract reference expression for every edge}
\usage{
extract_references(path_operations, milestone_network, reference_length = 100)
}
\arguments{
\item{path_operations}{The operations within each path}

\item{milestone_network}{The milestone network}

\item{reference_length}{The length in the reference of each population}
}
\description{
Extract reference expression for every edge
}
