% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{extract_variables}
\alias{extract_variables}
\title{Extract all variables from a formula}
\usage{
extract_variables(formula)
}
\arguments{
\item{formula}{An S4 class \code{AbstractFormula}}
}
\value{
A list of S4 class objects
}
\description{
Extract all variables from a formula
}
\examples{
extract_variables(fvar("c", 1) \%f*\% (fcon(1) - fvar("c", 2)))
}
