% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{\%f+\%}
\alias{\%f+\%}
\title{Addition function}
\usage{
lhs \%f+\% rhs
}
\arguments{
\item{lhs}{left hand side S4 class \code{AbstractFormula} object}

\item{rhs}{right hand side S4 class \code{AbstractFormula} object}
}
\value{
an S4 class \code{AbstractFormula} object
}
\description{
Addition function
}
\examples{
fvar("total") \%f+\% fcon(10)
}

