% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{fcon}
\alias{fcon}
\title{Create a new constant}
\usage{
fcon(value)
}
\arguments{
\item{value}{the value of the constant}
}
\value{
an S4 class \code{AbstractFormula} object
}
\description{
Create a new constant
}
\examples{
con1 <- fcon(1)
con2 <- fcon(2.1)
}

