% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_experiment.R
\name{filter_experiment}
\alias{filter_experiment}
\title{Filter counts}
\usage{
filter_experiment(experiment)
}
\arguments{
\item{experiment}{Experiment list, containing expression, cellinfo and geneinfo}
}
\description{
Filter counts
}
