% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_experiment.R
\name{filter_expression}
\alias{filter_expression}
\title{Filter expression}
\usage{
filter_expression(expression, cellinfo, geneinfo)
}
\arguments{
\item{expression}{Expression matrix}

\item{cellinfo}{Cell info dataframe}
}
\description{
Filter expression
}
