% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{finfix}
\alias{finfix}
\title{Create a new infix function to be used for the AbstractFormula system}
\usage{
finfix(fun)
}
\arguments{
\item{fun}{the function as a string}
}
\value{
an infix function which to create S4 class \code{AbstractFormula} objects
}
\description{
Create a new infix function to be used for the AbstractFormula system
}
\examples{
`\%f-\%` <- finfix("-")
}
