% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{fprefix}
\alias{fprefix}
\title{Create a new prefix function to be used for the AbstractFormula system}
\usage{
fprefix(fun)
}
\arguments{
\item{fun}{the function as a string}
}
\value{
a prefix function which to create S4 class \code{AbstractFormula} objects
}
\description{
Create a new prefix function to be used for the AbstractFormula system
}
\examples{
fsum <- fprefix("sum")
}
