% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{fprod}
\alias{fprod}
\title{Product of AbstractFormula objects}
\usage{
fprod(arguments)
}
\arguments{
\item{arguments}{the S4 class \code{AbstractFormula} objects to be multiplied}
}
\value{
an S4 class \code{AbstractFormula} object
}
\description{
Product of AbstractFormula objects
}
\examples{
fprod(lapply(c(1,2,3), fcon))
fprod(fcon(1), fcon(2))
}

