% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{freduce}
\alias{freduce}
\title{Apply a function over all arguments, then collapse it with another function}
\usage{
freduce(collapse.fun, lapply.args, lapply.fun)
}
\arguments{
\item{collapse.fun}{a function which can collapse multiple S4 class \code{AbstractFormula} objects into a new one}

\item{lapply.args}{arguments for the lapply}

\item{lapply.fun}{a function for the lapply that will return S4 class \code{AbstractFormula} objects}
}
\value{
an S4 class \code{AbstractFormula} object
}
\description{
Apply a function over all arguments, then collapse it with another function
}
\examples{
freduce(`\%f+\%`, c(1,2,3), fcon)
}
