% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b_random_tree.R
\name{from_stages_to_modulenet}
\alias{from_stages_to_modulenet}
\title{Convert a stage network to a module network which will}
\usage{
from_stages_to_modulenet(stagenet)
}
\description{
Convert a stage network to a module network which will
}
