% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{fsum}
\alias{fsum}
\title{Sum of AbstractFormula objects}
\usage{
fsum(...)
}
\arguments{
\item{...}{the S4 class \code{AbstractFormula} objects to be summed}
}
\value{
an S4 class \code{AbstractFormula} object
}
\description{
Sum of AbstractFormula objects
}
\examples{
fsum(lapply(c(1,2,3), fcon))
fsum(fcon(1), fcon(2))
}
