% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{fvar}
\alias{fvar}
\title{Create a new variable}
\usage{
fvar(prefix, ...)
}
\arguments{
\item{prefix}{the prefix of the variable}

\item{...}{optional arguments for the variable name}
}
\value{
an S4 class \code{AbstractFormula} object
}
\description{
Create a new variable
}
\examples{
total <- fvar("total")
gene <- fvar("gene", 1)
}
