% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barabasi_albert.R
\name{generate.ba}
\alias{generate.ba}
\title{Generate a Barabási–Albert graph}
\usage{
generate.ba(amnt.nodes, amnt.edges, reverse.edges = T, offset.exponent = 1,
  trace = F)
}
\arguments{
\item{amnt.nodes}{number of nodes in the generated graph}

\item{amnt.edges}{number of edges in the generated graph}

\item{reverse.edges}{if \code{TRUE} some nodes will have many outgoing links whereas others will not, 
else some nodes will have many incoming links whereas others will not}

\item{offset.exponent}{the exponent for the weighted sampling. A higher value will result in a 
smaller number of hub nodes having increasingly more outgoing edges.}

\item{trace}{will generate output in the console if \code{TRUE}}
}
\value{
a list containing several elements:
\itemize{
  \item amnt.nodes: the number of nodes in the graph
  \item degree: the degrees of each node
  \item in.degree: the degree of incoming edges of each node
  \item out.degree: the degree of outgoing edges of each node
  \item incoming.edges: a list of incoming edges for each of the nodes
  \item outgoing edges: a list of outgoing edges for each of the nodes
  \item data.frame: the edges in the form of a data frame
}
}
\description{
Generate a Barabási–Albert graph
}
\examples{
generate.ba(amnt.nodes = 100, amnt.edges = 1000, reverse.edges = T, offset.exponent = 1.5, trace = T)
}
