% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barabasi_albert_with_modules.R
\name{generate.ba.with.modules}
\alias{generate.ba.with.modules}
\title{Generate a Barabasi-Albert graph}
\usage{
generate.ba.with.modules(amnt.nodes, amnt.edges, exp1, exp2,
  reverse.edges = TRUE, trace = FALSE)
}
\arguments{
\item{amnt.nodes}{number of nodes in the generated graph}

\item{amnt.edges}{number of edges in the generated graph}

\item{exp1}{weight 1}

\item{exp2}{weight 2}

\item{reverse.edges}{if \code{TRUE} some nodes will have many outgoing links whereas others will not, 
else some nodes will have many incoming links whereas others will not}

\item{trace}{will generate output in the console if \code{TRUE}}
}
\value{
a list containing several elements:
\itemize{
  \item amnt.nodes: the number of nodes in the graph
  \item degree: the degrees of each node
  \item in.degree: the degree of incoming edges of each node
  \item out.degree: the degree of outgoing edges of each node
  \item incoming.edges: a list of incoming edges for each of the nodes
  \item outgoing edges: a list of outgoing edges for each of the nodes
  \item data.frame: the edges in the form of a data frame
}
}
\description{
Generate a Barabasi-Albert graph
}
\examples{
generate.ba.with.modules(
  amnt.nodes = 100, 
  amnt.edges = 1000, 
  reverse.edges = TRUE, 
  exp1 = 1.5, 
  exp2 = 1.5, 
  trace = TRUE
)
}
