% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{generate.decay.formula}
\alias{generate.decay.formula}
\title{Generate decay formulae}
\usage{
generate.decay.formula(g, amnt.genes)
}
\arguments{
\item{g}{the gene in question}

\item{amnt.genes}{the total number of genes}
}

