% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{generate.production.formula}
\alias{generate.production.formula}
\title{Generate production formulae}
\usage{
generate.production.formula(g, regs, amnt.genes)
}
\arguments{
\item{g}{the gene in question}

\item{regs}{the regulators of the gene}

\item{amnt.genes}{the total number of genes}
}
