% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy-3-expression.R
\name{generate_counts}
\alias{generate_counts}
\title{Generate counts from expression data}
\usage{
generate_counts(expression)
}
\arguments{
\item{expression}{The original expression data}
}
\description{
Generate counts from expression data
}
