% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy-3-expression.R
\name{generate_expression}
\alias{generate_expression}
\title{Generate toy expression based on a milestone_network and cell progression information}
\usage{
generate_expression(milestone_network, progressions, ngenes = 100,
  noise_std = 0.05)
}
\arguments{
\item{milestone_network}{The milestone network}

\item{progressions}{The progressions}

\item{ngenes}{The number of genes to generate}

\item{noise_std}{Noise parameter}
}
\description{
Generate toy expression based on a milestone_network and cell progression information
}
