% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_feature_network.R
\name{generate_feature_network}
\alias{generate_feature_network}
\alias{feature_network_default}
\alias{.feature_network_random_realnet}
\title{Generate a target network}
\usage{
generate_feature_network(model)

feature_network_default(realnet = .feature_network_random_realnet(),
  damping = 0.05, target_resampling = Inf, max_in_degree = 5)

.feature_network_random_realnet()
}
\arguments{
\item{model}{A dyngen intermediary model for which the transcription network has been generated with \code{\link[=generate_tf_network]{generate_tf_network()}}.}

\item{realnet}{The name of a gene regulatory network (GRN) in \link{realnets}.
Alternatively, a custom GRN can be used by passing a weighted sparse matrix.}

\item{damping}{A damping factor used for the page rank algorithm used to subsample the realnet.}

\item{target_resampling}{How many targets / HKs to sample from the realnet per iteration.}

\item{max_in_degree}{The maximum in-degree of a target / HK.}
}
\description{
\code{\link[=generate_feature_network]{generate_feature_network()}} generates a network of target genes that are regulated
by the previously generated TFs, and also a separate network of housekeeping genes (HKs).
\code{\link[=feature_network_default]{feature_network_default()}} is used to configure parameters pertaining this process.
}
