% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_formulae.R
\name{generate_formulae}
\alias{generate_formulae}
\title{Generate formulae from a given network}
\usage{
generate_formulae(net, geneinfo, cells = tibble(cell_id = 1, dies = FALSE))
}
\arguments{
\item{net}{Network dataframe}

\item{geneinfo}{Gene info dataframe}

\item{cells}{Cells dataframe}
}
\description{
Generate formulae from a given network
}
