% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_kinetics.R
\name{generate_kinetics}
\alias{generate_kinetics}
\alias{kinetics_default}
\title{Determine the kinetics of the feature network}
\usage{
generate_kinetics(model)

kinetics_default()
}
\arguments{
\item{model}{A dyngen intermediary model for which the feature network has been generated with \code{\link[=generate_feature_network]{generate_feature_network()}}.}
}
\description{
\code{\link[=generate_kinetics]{generate_kinetics()}} samples the kinetics of genes in the feature network for which
the kinetics have not yet been defined.
\code{\link[=kinetics_default]{kinetics_default()}} is used to configure parameters pertaining this process.
}
