% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b_random_tree.R
\name{generate_random_tree}
\alias{generate_random_tree}
\title{Generate a random tree}
\usage{
generate_random_tree(treeseed = NULL, decay = 1.4)
}
\arguments{
\item{treeseed}{A seed to set}

\item{decay}{The decay rate}
}
\description{
Generate a random tree
}
