% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_kinetics.R
\name{randomize_gene_kinetics}
\alias{randomize_gene_kinetics}
\alias{randomize_cell_kinetics}
\alias{generate_system}
\title{Randomize kinetics of genes}
\usage{
randomize_gene_kinetics(feature_info, net,
  samplers = get_default_kinetics_samplers())

randomize_cell_kinetics(cells)

generate_system(net, feature_info, cells,
  samplers = get_default_kinetics_samplers())
}
\arguments{
\item{feature_info}{Geneinfo dataframe}

\item{net}{Regulatory network dataframe}

\item{samplers}{The samplers for the kinetics parameters}

\item{cells}{Cells dataframe}
}
\description{
Randomize kinetics of genes

Randomize kinetics of cells

Generate a system from a network using
}
