% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy-1-networks.R
\name{generate_toy_milestone_network}
\alias{generate_toy_milestone_network}
\title{Generate a toy milestone network}
\usage{
generate_toy_milestone_network(ti_type = c("linear", "bifurcating", "cycle"))
}
\arguments{
\item{ti_type}{The type of network to generate. Accepted: linear, bifurcating, cycle.}
}
\description{
Generate a toy milestone network
}
