% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b_random_tree.R
\name{generate_tree_states}
\alias{generate_tree_states}
\title{Convert tree to states list, used later by the gold standard to determine where there are bifurcations
The modules of each bifurcation will be added later by `from_stages_to_modulenet`}
\usage{
generate_tree_states(stagenet)
}
\arguments{
\item{stagenet}{The state network to be converted.}
}
\description{
Convert tree to states list, used later by the gold standard to determine where there are bifurcations
The modules of each bifurcation will be added later by `from_stages_to_modulenet`
}
