% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_experiment.R
\name{get_housekeeping_reference_means}
\alias{get_housekeeping_reference_means}
\title{Get housekeeping reference means}
\usage{
get_housekeeping_reference_means()
}
\arguments{
\item{counts}{Expression matrix containing counts}
}
\description{
Get housekeeping reference means
}
