% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_goldstandard.R
\name{get_milestone_paths}
\alias{get_milestone_paths}
\title{Extract all milestone paths}
\usage{
get_milestone_paths(milestone_network, start_milestones, max_path_length = 10)
}
\arguments{
\item{milestone_network}{The milestone network}

\item{start_milestones}{The starting milestones}

\item{max_path_length}{Maximal number of milestones per path}
}
\description{
Extract all milestone paths
}
