% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timings.R
\name{get_timings}
\alias{get_timings}
\title{Return the timings of each of the dyngen steps}
\usage{
get_timings(model)
}
\arguments{
\item{model}{A dyngen object}
}
\description{
Return the timings of each of the dyngen steps
}
\examples{
model <- 
  initialise_model(backbone = backbone_linear()) \%>\% 
  generate_tf_network()

timings <- get_timings(model)
}
