% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ginhoux}
\alias{ginhoux}
\title{scRNA-seq data of dendritic cell progenitors.}
\format{A list containing two data frames, \code{expression} and \code{sample_info}.}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE60783}
}
\usage{
ginhoux
}
\description{
This dataset contains the expression values for 248 dendritic cell progenitors.
Each cell is in one of three maturation stages: MDP, CDP or PreDC.
The levels of the factor in \code{sample.info} are ordered according to the maturation process.
}
\references{
Schlitzer A, Sivakamasundari V, Chen J, Sumatoh HR et al.
Identification of cDC1- and cDC2-committed DC progenitors reveals early lineage priming at
 the common DC progenitor stage in the bone marrow. Nat Immunol 2015 Jul;16(7):718-28. PMID: 26054720
}
\keyword{datasets}
