% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_simulation.R
\name{kinetics_noise_none}
\alias{kinetics_noise_none}
\alias{kinetics_noise_simple}
\title{Add small noise to the kinetics of each simulation}
\usage{
kinetics_noise_none()

kinetics_noise_simple(mean = 1, sd = 0.005)
}
\arguments{
\item{mean}{The mean level of noise (should be 1)}

\item{sd}{The sd of the noise (should be a relatively small value)}
}
\description{
Add small noise to the kinetics of each simulation
}
