% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_loading_saving.R
\name{load_dataset}
\alias{load_dataset}
\title{Load a dataset}
\usage{
load_dataset(datasetid, contents = contents_dataset())
}
\arguments{
\item{datasetid}{The id of the dataset}

\item{contents}{What to load}
}
\description{
Load a dataset
}
