% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_wraptasks.R
\name{load_datasets}
\alias{load_datasets}
\title{Load dyngen datasets}
\usage{
load_datasets(datasets_info = load_datasets_info(), mc_cores = 1)
}
\arguments{
\item{datasets_info}{The meta information of the datasets to read}

\item{mc_cores}{The number of cores to use while loading in the datasets (default: 1)}
}
\value{
A tibble of datasets
}
\description{
\strong{Important!} A variable \code{.datasets_location} needs to be defined in the global environment for this function
to work correctly.
}
\examples{
\dontrun{
.dataset_location <- "path_to_dyngen_results"
datasets <- load_datasets(mc_cores = 1, datasets_info = load_datasets_info())
}
}
