% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_wraptasks.R
\name{load_datasets_info}
\alias{load_datasets_info}
\title{Load dyngen datasets meta information}
\usage{
load_datasets_info()
}
\value{
The meta information of the datasets
}
\description{
\strong{Important!} A variable \code{.datasets_location} needs to be defined in the global environment for this function
to work correctly.
}
\examples{
\dontrun{
.dataset_location <- "path_to_dyngen_results"
load_datasets_info()
}
}
