% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_network_generation.R
\name{load_modulenet}
\alias{load_modulenet}
\alias{list_modulenets}
\title{Load module network from extra data}
\usage{
load_modulenet(modulenet_name)

list_modulenets()
}
\arguments{
\item{modulenet_name}{A modulenet_name as defined in
\href{https://github.com/dynverse/dyngen/tree/master/inst/ext_data/modulenetworks}{ext_data/modulenetworks}.
Additionally, \code{"tree"} can also be used in order to generate a random tree.}
}
\description{
Load module network from extra data
}
