% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_kinetics.R
\name{merge_models}
\alias{merge_models}
\title{Merge different models}
\usage{
merge_models(model1, model2)
}
\arguments{
\item{model1}{The first model}

\item{model2}{The second model}
}
\description{
Merge different models
}
