% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1a_module_networks.R
\name{module_network}
\alias{module_network}
\alias{module_network_bifurcating}
\alias{module_network_bifurcating_converging}
\alias{module_network_bifurcating_cycle}
\alias{module_network_bifurcating_loop}
\alias{module_network_binary_tree}
\alias{module_network_consecutive_bifurcating}
\alias{module_network_converging}
\alias{module_network_cycle}
\alias{module_network_linear}
\alias{module_network_linear_long}
\alias{module_network_trifurcating}
\title{Module network models for starting the simulation with}
\usage{
module_network(nodes, edges, operations)

module_network_bifurcating()

module_network_bifurcating_converging()

module_network_bifurcating_cycle()

module_network_bifurcating_loop()

module_network_binary_tree()

module_network_consecutive_bifurcating()

module_network_converging()

module_network_cycle()

module_network_linear()

module_network_linear_long()

module_network_trifurcating()
}
\description{
A module is a group of genes which, to some extent, shows the same
expression behaviour. Several modules are connected together such that
one or more genes from one module will regulate the expression of
another module. By creating chains of modules, a dynamic behaviour in gene
regulation can be created.
}
