% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_generate_modulenet.R
\name{modulenet}
\alias{modulenet}
\alias{list_modulenet_generators}
\alias{generate_modulenet_bifurcating}
\alias{generate_modulenet_bifurcating_converging}
\alias{generate_modulenet_bifurcating_cycle}
\alias{generate_modulenet_bifurcating_loop}
\alias{generate_modulenet_binary_tree}
\alias{generate_modulenet_consecutive_bifurcating}
\alias{generate_modulenet_converging}
\alias{generate_modulenet_cycle}
\alias{generate_modulenet_linear}
\alias{generate_modulenet_linear_long}
\alias{generate_modulenet_trifurcating}
\title{Module network models for starting the simulation with}
\usage{
modulenet(module_nodes, module_module_edges, expression_patterns)

list_modulenet_generators()

generate_modulenet_bifurcating()

generate_modulenet_bifurcating_converging()

generate_modulenet_bifurcating_cycle()

generate_modulenet_bifurcating_loop()

generate_modulenet_binary_tree()

generate_modulenet_consecutive_bifurcating()

generate_modulenet_converging()

generate_modulenet_cycle()

generate_modulenet_linear()

generate_modulenet_linear_long()

generate_modulenet_trifurcating()
}
\description{
A module is a group of genes which, to some extent, shows the same
expression behaviour. Several modules are connected together such that
one or more genes from one module will regulate the expression of
another module. By creating chains of modules, a dynamic behaviour in gene
regulation can be created.
}
\details{
A modulenet model contains three tibbles, namely \code{module_nodes},
\code{module_edges}, and \code{expression_patterns}.

\code{module_nodes} contains meta information on the modules themselves.
\itemize{
\item module_id (character): the name of the module
\item a0 (numeric): basal expression level of genes in this module
\item burn (logical): whether or not outgoing edges of this module will
be active during the burn in phase
}

\code{module_edges} describes which modules regulate which other modules.
\itemize{
\item from (character): the regulating module
\item to (character): the target module
\item effect (integer): \code{1} if the regulating module upregulates
the target module, \code{-1} if it downregulates
\item strength (numeric): the strength of the interaction
\item cooperativity (numeric): what does this do again?
}

\code{expression_patterns} describes the expected expression pattern
changes when a cell is simulated by dyngen.
\itemize{
\item from (character): name of a cell state
\item to (character): name of a cell state
\item module_progression (character): differences in module expression between the two states.
Example: \code{"+4,-1|+9|-4"} means the expression of module 4 will go up at the same time as module 1 goes down;
afterwards module 9 expression will go up, and afterwards module 4 expression will go down again.
\item start (logical): Whether or not this from cell state is the start of the trajectory
\item burn (logical): Whether these cell states are part of the burn in phase. Cells will
not get sampled from these cell states.
}
}
