% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_network_generation.R
\name{modulenet_to_genenet}
\alias{modulenet_to_genenet}
\title{Convert modulenet to modules regulating each other}
\usage{
modulenet_to_genenet(modulenet, modulenodes, ngenes_per_module = function(n)
  sample(1:10, n, replace = TRUE), gene_name_generator = function(i)
  paste0("GM", i), edge_retainment = function(n) max(c(round(n/2), 1)))
}
\arguments{
\item{modulenet}{Module network}

\item{modulenodes}{Module nodes}

\item{edge_retainment}{Function for sampling the number of edges retained between tfs of module nodes}

\item{ngenes_per_modules}{Functions for sampling the number of genes per module}

\item{gene_name_generate}{FUnction for generating the name of a gene}
}
\description{
Convert modulenet to modules regulating each other
}
