% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_loading_saving.R
\name{overviewer}
\alias{overviewer}
\title{Generate an overview data.frame for a particular type of data}
\usage{
overviewer(type)
}
\arguments{
\item{type}{The type of data}
}
\description{
Generate an overview data.frame for a particular type of data
}
