% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_platform.R
\name{platform_from_counts}
\alias{platform_from_counts}
\alias{platform_simple}
\title{Estimate platform parameters from a dataset}
\usage{
platform_from_counts(counts, grouping, subsample = 500)

platform_simple(n_cells = 100L, n_features = 100L,
  trajectory_dependent_features = 0.1, dropout_rate = 0.01,
  dropout_shape = 1)
}
\arguments{
\item{counts}{The counts with cells in rows and genes in columns.}

\item{grouping}{A named vector representing a grouping of the cells.}

\item{subsample}{The number of cells to subsample.}

\item{n_cells}{The number of cells}

\item{n_features}{The number of features}

\item{trajectory_dependent_features}{The percentage of features that are being driven by the trajectory (or vice versa)}

\item{dropout_mean_rate}{The mean rate of dropouts}

\item{dropout_mean_shape}{The shape of dropouts}
}
\description{
Altenatively, \code{\link[=platform_simple]{platform_simple()}} returns a toy platform parameter configuration.
}
