% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_backbone}
\alias{plot_backbone}
\title{Visualise the backbone of a model}
\usage{
plot_backbone(model)
}
\arguments{
\item{model}{A dyngen initial model created with \code{\link[=initialise_model]{initialise_model()}}.}
}
\description{
Visualise the backbone of a model
}
