% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_backbone_statenet}
\alias{plot_backbone_statenet}
\title{Visualise the backbone state network of a model}
\usage{
plot_backbone_statenet(model, detailed = FALSE)
}
\arguments{
\item{model}{A dyngen initial model created with \code{\link[=initialise_model]{initialise_model()}}.}

\item{detailed}{Whether or not to also plot the substates of transitions.}
}
\description{
Visualise the backbone state network of a model
}
