% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_experiment}
\alias{plot_experiment}
\title{Plotting functions for experiment}
\usage{
plot_experiment(experiment)
}
\arguments{
\item{experiment}{The experiment}
}
\description{
Plotting functions for experiment
}
