% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_experiment_dimred}
\alias{plot_experiment_dimred}
\title{Plot a dimensionality reduction of the final dataset}
\usage{
plot_experiment_dimred(model, mapping = aes_string("comp_1", "comp_2"))
}
\arguments{
\item{model}{A dyngen intermediary model for which the simulations have been run with \code{\link[=generate_experiment]{generate_experiment()}}.}

\item{mapping}{Which components to plot.}
}
\value{
A ggplot2 object.
}
\description{
Plot a dimensionality reduction of the final dataset
}
\examples{
\donttest{
data("example_model")
plot_experiment_dimred(example_model)
}
}
