% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gold_simulations_proj}
\alias{plot_gold_simulations_proj}
\title{Visualise the simulations using mRNA and pre-mRNA information}
\usage{
plot_gold_simulations_proj(model, detailed = FALSE,
  mapping = aes(comp_1, comp_2))
}
\arguments{
\item{model}{A dyngen intermediary model for which the simulations have been run with \code{\link[=generate_cells]{generate_cells()}}.}

\item{detailed}{Whether or not to colour according to each separate sub-edge in the gold standard.}

\item{mapping}{Which components to plot.}
}
\description{
Visualise the simulations using mRNA and pre-mRNA information
}
