% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_model}
\alias{plot_model}
\alias{plot_module_network}
\alias{plot_feature_network}
\title{Model plotting}
\usage{
plot_model(model)

plot_module_network(model)

plot_feature_network(model, color_by = c("module", "main"),
  tfs_only = FALSE, label = FALSE)
}
\arguments{
\item{model}{The model}

\item{color_by}{By what to color}

\item{label}{Whether to label genes}

\item{main_only}{Whether to only draw the main network}
}
\description{
Model plotting
}
