% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_net}
\alias{plot_net}
\title{Plot a gene network}
\usage{
plot_net(model, colorby = c("module", "main"), main_only = TRUE)
}
\arguments{
\item{model}{The model}

\item{colorby}{By what to color}

\item{main_only}{Whether to only draw the main network}
}
\description{
Plot a gene network
}
