% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_net_overlaps}
\alias{plot_net_overlaps}
\title{Plot the overlap in targets in a heatmap}
\usage{
plot_net_overlaps(model)
}
\arguments{
\item{model}{The model}
}
\description{
Plot the overlap in targets in a heatmap
}
