% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_normalisation}
\alias{plot_normalisation}
\title{Plotting functions for normalisation}
\usage{
plot_normalisation(experiment, normalisation)
}
\arguments{
\item{experiment}{The experiment}

\item{normalisation}{The normalisation}
}
\description{
Plotting functions for normalisation
}
