% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_simulation}
\alias{plot_simulation}
\alias{plot_simulation_space_individual}
\alias{plot_simulation_space_time}
\alias{plot_simulation_space_modules}
\alias{plot_simulation_simulations_lines}
\alias{plot_simulation_modules_lines}
\alias{plot_simulation_modules_heatmap}
\alias{plot_simulation_3D}
\title{Simulation plotting}
\usage{
plot_simulation(simulation)

plot_simulation_space_individual(simulation,
  spaces = dimred_simulation(simulation))

plot_simulation_space_time(simulation, spaces = dimred_simulation(simulation))

plot_simulation_space_modules(simulation,
  subsample = subsample_simulation(simulation),
  spaces = dimred_simulation(simulation, subsample, "ica",
  "samplexpression_modules"))

plot_simulation_simulations_lines(simulation)

plot_simulation_modules_lines(simulation)

plot_simulation_modules_heatmap(simulation)

plot_simulation_3D(simulation)
}
\arguments{
\item{simulation}{The simulation}

\item{spaces}{Spaces dataframe}

\item{subsample}{Subsampled expression data}
}
\description{
Simulation plotting
}
