% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_simulations}
\alias{plot_simulations}
\title{Visualise the simulations using the dimred}
\usage{
plot_simulations(model, mapping = aes(comp_1, comp_2))
}
\arguments{
\item{model}{A dyngen intermediary model for which the simulations have been run with \code{\link[=generate_cells]{generate_cells()}}.}

\item{mapping}{Which components to plot.}
}
\description{
Visualise the simulations using the dimred
}
