% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_gs.R
\name{preprocess_simulation_for_gs}
\alias{preprocess_simulation_for_gs}
\title{Preprocessing of simulation prior to gold standard determination
1. Smoothing of expression}
\usage{
preprocess_simulation_for_gs(simulation, model)
}
\description{
Preprocessing of simulation prior to gold standard determination
1. Smoothing of expression
}
