% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy-2-cells.R
\name{random_progressions}
\alias{random_progressions}
\title{Assign random progressions without tenting}
\usage{
random_progressions(milestone_network, ncells = 100)
}
\description{
Assign random progressions without tenting
}
