% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_kinetics.R
\name{randomize_cell_kinetics}
\alias{randomize_cell_kinetics}
\title{Randomize kinetics of cells}
\usage{
randomize_cell_kinetics(cells)
}
\arguments{
\item{cells}{Cells dataframe}
}
\description{
Randomize kinetics of cells
}
