% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_kinetics.R
\name{randomize_gene_kinetics}
\alias{randomize_gene_kinetics}
\title{Randomize kinetics of genes}
\usage{
randomize_gene_kinetics(geneinfo, net, samplers = list(sample_r = function(n)
  runif(n, 10, 200), sample_d = function(n) runif(n, 2, 8), sample_p =
  function(n) runif(n, 2, 8), sample_q = function(n) runif(n, 1, 5),
  calculate_a0 = function(effects) {     if (length(effects) == 0) {         1  
    } else if (sum(effects == -1) == 0) {         1e-04     } else if
  (sum(effects == 1) == 0) {         1     } else {         0.5     } },
  calculate_a = function(configuration_id, effects) {     bound <-
  get_binding_configuration(configuration_id, length(effects))     if
  (any(effects[bound] == -1)) {         0     } else {         1     } },
  sample_strength = function(n) runif(n, 1, 20), calculate_k =
  function(max_protein, strength) {     max_protein/2/strength },
  sample_cooperativity = function(n) runif(n, 1, 4)))
}
\arguments{
\item{geneinfo}{Geneinfo dataframe}

\item{net}{Regulatory network dataframe}
}
\description{
Randomize kinetics of genes
}
