% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-realnet.R
\docType{data}
\name{realcounts}
\alias{realcounts}
\title{A set of real single cell expression datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 110 rows and 2 columns.
}
\usage{
realcounts
}
\description{
Statistics are derived from these datasets in order to simulate single cell experiments.
}
\keyword{datasets}
