% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-realnet.R
\docType{data}
\name{realnets}
\alias{realnets}
\title{A set of gold standard gene regulatory networks}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 2 columns.}
\usage{
realnets
}
\description{
These networks are subsampled in order to generate realistic feature and housekeeping networks.
}
\keyword{datasets}
