% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ssa_etl}
\alias{ssa_exact}
\alias{ode_em}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{GillespieSSA2}{\code{\link[GillespieSSA2]{ode_em}}, \code{\link[GillespieSSA2]{ssa_etl}}, \code{\link[GillespieSSA2]{ssa_exact}}}

  \item{purrr}{\code{\link[purrr:pipe]{\%>\%}}}
}}

