% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm_bounded.R
\name{rnorm_bounded}
\alias{rnorm_bounded}
\title{A bounded version of rnorm}
\usage{
rnorm_bounded(n, mean = 0, sd = 1, min = -Inf, max = Inf)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}

\item{min}{lower limits of the distribution.}

\item{max}{upper limits of the distribution.}
}
\description{
A bounded version of rnorm
}
\examples{
rnorm_bounded(10)
}
