% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_loading_saving.R
\name{saver}
\alias{saver}
\title{Save a dataset of a particular type}
\usage{
saver(x, type, overview_only = FALSE)
}
\arguments{
\item{x}{?? TODO}

\item{type}{?? TODO}

\item{overview_only}{?? TODO}
}
\description{
Save a dataset of a particular type
}
