% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_simulate.R
\name{simulate_cell}
\alias{simulate_cell}
\title{Simulation one individual cells}
\usage{
simulate_cell(system, timeofsampling = NULL, totaltime = 10, burntime = 2,
  ssa_algorithm = fastgssa::ssa.em(noise_strength = 4))
}
\arguments{
\item{system}{The system to simulate}

\item{timeofsampling}{`NULL` to return expression at all time steps, a double to return expression at a particular time step}

\item{totaltime}{The total simulation time}

\item{burntime}{A burnin time}

\item{ssa_algorithm}{Which SSA algorithm to use}
}
\description{
Simulation one individual cells
}
