% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_simulate.R
\name{simulate_multiple}
\alias{simulate_multiple}
\title{Simulate multiple cells}
\usage{
simulate_multiple(system, burntime, totaltime, nsimulations = 16,
  ssa_algorithm = fastgssa::ssa.em(noise_strength = 4))
}
\arguments{
\item{system}{The system to simulate}

\item{burntime}{The burn-in time before sampling}

\item{totaltime}{The total simulation time}

\item{nsimulations}{The number of simulations}

\item{ssa_algorithm}{Which GSSA algorithm to use}
}
\description{
Simulate multiple cells
}
