% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_experiment.R
\name{snapshot_sampler}
\alias{snapshot_sampler}
\title{Snapshot sampler}
\usage{
snapshot_sampler(weight_bw = 0.1)
}
\arguments{
\item{weight_bw}{The bandwidth for density estimation, measured in "percentage" units}
}
\description{
Snapshot sampler
}
