% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastgssa.R
\name{ssa_direct}
\alias{ssa_direct}
\title{Direct method (D)}
\usage{
ssa_direct()
}
\value{
an object of to be used by \code{\link{ssa}}.
}
\description{
Direct method implementation of the \acronym{SSA} as described by Gillespie (1977).
}
