% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_experiment.R
\name{synchronised_sampler}
\alias{synchronised_sampler}
\title{Snapshot sampler}
\usage{
synchronised_sampler(ntimepoints = 10)
}
\arguments{
\item{ntimepoints}{Number of timepoints to sample}
}
\description{
Snapshot sampler
}
