% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_experiment.R
\name{take_experiment_cells}
\alias{take_experiment_cells}
\title{Extract cells from a simulation, either simulating a snapshot experiment or a synchronized experiment}
\usage{
take_experiment_cells(simulation, model, samplesettings = list(type =
  "snapshot", ncells = 500))
}
\arguments{
\item{simulation}{List containing the expression of simulation(s)}

\item{model}{The model of the simulation}

\item{samplesettings}{List containing the information of how the cells should be sampled}
}
\description{
Extract cells from a simulation, either simulating a snapshot experiment or a synchronized experiment
}
