% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_genenet.R
\name{target_realnet}
\alias{target_realnet}
\title{... todo}
\usage{
target_realnet(realnet_name = realnets$name, damping = 0.05,
  ntargets_sampler = function(n_features, n_regulators) sample(20:100,
  1))
}
\description{
... todo
}
