% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_targetadder.R
\name{targetadder_realnet}
\alias{targetadder_realnet}
\title{... todo}
\usage{
targetadder_realnet(realnet_name = realnets$name, damping = 0.05,
  ntargets_sampler = function(n_features, n_regulators) sample(20:100,
  1), gene_name_generator = function(i) paste0("G", i))
}
\description{
... todo
}
