% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_dynwrap.R
\name{wrap_dataset}
\alias{wrap_dataset}
\title{Wrap a simulation into a dynwrap object}
\usage{
wrap_dataset(model)
}
\arguments{
\item{model}{A dyngen output model for which the experiment has been emulated with \code{\link[=generate_experiment]{generate_experiment()}}.}
}
\description{
The output of this object can be used with \pkg{dyno}.
}
