% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_wraptasks.R
\name{wrap_task}
\alias{wrap_task}
\title{Wrap tasks}
\usage{
wrap_task(id = "", params, model, simulation, gs, experiment, normalisation)
}
\arguments{
\item{params}{Parameters of the dataset}

\item{model}{Model}

\item{simulation}{Simulation}

\item{gs}{Gold standard}

\item{experiment}{Experiment}

\item{normalisation}{Normalisation}
}
\description{
Wrap tasks
}
