% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_simulate.R
\name{add_housekeeping_poisson}
\alias{add_housekeeping_poisson}
\title{Add housekeeping genes}
\usage{
add_housekeeping_poisson(expression, geneinfo, ngenes = 200,
  overallaverage = mean(expression))
}
\description{
Add housekeeping genes
}
