% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_simulate.R
\name{add_housekeeping_poisson}
\alias{add_housekeeping_poisson}
\title{Add housekeeping genes}
\usage{
add_housekeeping_poisson(expression, geneinfo, ngenes = 200,
  overallaverage = mean(expression))
}
\arguments{
\item{expression}{The original expression data.}

\item{geneinfo}{The original gene info}

\item{ngenes}{The number of genes to add}

\item{overallaverage}{TODO: Zouter/wouters help?}
}
\description{
Add housekeeping genes
}
