% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_experiment.R
\name{add_housekeeping_poisson}
\alias{add_housekeeping_poisson}
\title{Add housekeeping genes}
\usage{
add_housekeeping_poisson(expression, geneinfo, housekeeping_reference_means,
  n_housekeeping_genes = 200, overallaverage = mean(expression),
  gene_id_generator = function(n) {     paste0("GH", seq_len(n)) })
}
\arguments{
\item{expression}{The original expression data.}

\item{geneinfo}{The original gene info}

\item{overallaverage}{Overall average expression of the original dataset, this keeps the overall average expression the same even with housekeeping genes}

\item{ngenes}{The number of genes to add}

\item{meanpoissons}{The mean expression of a set of genes in the reference dataset}
}
\description{
Add housekeeping genes
}
