% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{dyngen}
\alias{dyngen}
\title{dyngen: A multi-modal simulator for spearheading single-cell omics analyses}
\description{
A toolkit for generating synthetic single cell data.
}
\examples{
model <- initialise_model(
  backbone = backbone_bifurcating(),
  # set this to the number of cores in your system
  num_cores = 4, 
  # set this to a directory where dyngen can cache some files
  download_cache_dir = "~/.cache/dyngen" 
)
\dontshow{
# actually use a smaller example 
# to reduce execution time during
# testing of the examples
model <- initialise_model(
  backbone = model$backbone,
  num_cells = 5,
  num_targets = 0,
  num_hks = 0,
  gold_standard_params = gold_standard_default(census_interval = 1, tau = 0.1),
  simulation_params = simulation_default(
    burn_time = 10,
    total_time = 10,
    census_interval = 1,
    ssa_algorithm = ssa_etl(tau = 0.1),
    experiment_params = simulation_type_wild_type(num_simulations = 1)
  )
)
}
\donttest{
model <- model \%>\%
  generate_tf_network() \%>\%
  generate_feature_network() \%>\%
  generate_kinetics() \%>\%
  generate_gold_standard() \%>\%
  generate_cells() \%>\%
  generate_experiment()
  
dataset <- wrap_dataset(model)

# dynplot::plot_dimred(dataset)
}
}
