% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b_random_tree.R
\name{from_stages_to_modulenet}
\alias{from_stages_to_modulenet}
\title{Convert a stage network to a module network}
\usage{
from_stages_to_modulenet(stagenet)
}
\arguments{
\item{stagenet}{A network of stages}
}
\description{
Convert a stage network to a module network
}
