% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_simulation.R
\name{generate_cells}
\alias{generate_cells}
\alias{simulation_default}
\title{Simulate the cells}
\usage{
generate_cells(model)

simulation_default(burn_time = 2, total_time = 10,
  ssa_algorithm = ssa_etl(tau = 0.001), census_interval = 0.01,
  num_simulations = 32, seeds = sample.int(10 * num_simulations,
  num_simulations))
}
\arguments{
\item{model}{A dyngen intermediary model for which the gold standard been generated with \code{\link[=generate_gold_standard]{generate_gold_standard()}}.}

\item{burn_time}{The burn in time of the system, used to determine an initial state vector.}

\item{total_time}{The total simulation time of the system.}

\item{ssa_algorithm}{Which SSA algorithm to use for simulating the cells with \code{\link[GillespieSSA2:ssa]{GillespieSSA2::ssa()}}}

\item{census_interval}{A granularity parameter for the outputted simulation.}

\item{num_simulations}{The number of simulations to run.}

\item{seeds}{A set of seeds for each of the simulations.}
}
\description{
\code{\link[=generate_cells]{generate_cells()}} runs simulations in order to determine the gold standard
of the simulations.
\code{\link[=simulation_default]{simulation_default()}} is used to configure parameters pertaining this process.
}
