% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_kinetics.R
\name{generate_kinetics}
\alias{generate_kinetics}
\alias{kinetics_default}
\title{Determine the kinetics of the feature network}
\usage{
generate_kinetics(model)

kinetics_default(sample_wpr = function(n) rnorm(n, 100, 20) \%>\%
  pmax(10), sample_wdr = function(n) rnorm(n, 5, 1) \%>\% pmax(2),
  sample_xpr = function(n) rnorm(n, 25, 5) \%>\% pmax(5),
  sample_xdr = function(n) rnorm(n, 5, 1) \%>\% pmax(2),
  sample_ypr = function(n) rnorm(n, 5, 1) \%>\% pmax(2),
  sample_ydr = function(n) rnorm(n, 3, 0.5) \%>\% pmax(1),
  sample_effect = function(n) sample(c(-1, 1), n, replace = TRUE, prob =
  c(0.25, 0.75)), sample_strength = function(n) 10^runif(n, log10(1),
  log10(100)), sample_cooperativity = function(n) runif(n, 0.5, 2))
}
\arguments{
\item{model}{A dyngen intermediary model for which the feature network has been generated with \code{\link[=generate_feature_network]{generate_feature_network()}}.}

\item{sample_wpr}{A function specifying the distribution from which to sample the pre-mRNA production rate.}

\item{sample_wdr}{A function specifying the distribution from which to sample the pre-mRNA decay rate.}

\item{sample_xpr}{A function specifying the distribution from which to sample the mRNA production rate.}

\item{sample_xdr}{A function specifying the distribution from which to sample the mRNA decay rate.}

\item{sample_ypr}{A function specifying the distribution from which to sample the protein production rate.}

\item{sample_ydr}{A function specifying the distribution from which to sample the protein decay rate.}

\item{sample_effect}{A function specifying the distribution from which to sample the effect of an interaction.}

\item{sample_strength}{A function specifing the distribution from which to sample the strength of an interaction.}

\item{sample_cooperativity}{A function specifing the distribution from which to sample the cooperativity of an interaction from.}
}
\description{
\code{\link[=generate_kinetics]{generate_kinetics()}} samples the kinetics of genes in the feature network for which
the kinetics have not yet been defined.
\code{\link[=kinetics_default]{kinetics_default()}} is used to configure parameters pertaining this process.
}
