% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_simulation.R
\name{generate_model_from_modulenet}
\alias{generate_model_from_modulenet}
\title{Model initializations}
\usage{
generate_model_from_modulenet(model)
}
\arguments{
\item{model}{the module model -- is a list which must contain modulenodes, modulenet, celltypes and states}
}
\description{
Model initializations
}
