% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_network_generation.R
\name{generate_model_from_modulenet}
\alias{generate_model_from_modulenet}
\title{Model generation from modulenet}
\usage{
generate_model_from_modulenet(modulenet_name, treeseed = NULL, decay = 1.4,
  ngenes_per_module = function(n) sample(1:10, n, replace = TRUE),
  edge_retainment = function(n) max(c(round(n/2), 1)),
  ntargets_sampler = function() {     sample(20:100, 1) }, target_adder_name,
  realnet_name = "regulatorycircuits", damping = 0.05, samplers)
}
\arguments{
\item{modulenet_name}{A modulenet_name as defined in data/modulenetworks or `tree` to generate a random tree}

\item{treeseed}{A seed to set}

\item{decay}{The decay rate}

\item{ngenes_per_module}{Functions for sampling the number of genes per module}

\item{edge_retainment}{Function for sampling the number of edges retained between tfs of module nodes}

\item{ntargets_sampler}{A sample function for the number of targets}

\item{target_adder_name}{The method for adding targets. Only "realnet" is currently supported.}

\item{realnet_name}{Name of the real network.}

\item{damping}{A daping factor for personalized pagerank}

\item{samplers}{The samplers for the kinetics parameters}
}
\description{
Default parameters are defined in individual functions and are inherited in `zzz_param_inheritance.R`
}
