% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_network_generation.R
\name{generate_model_from_modulenet}
\alias{generate_model_from_modulenet}
\title{Model generation from modulenet}
\usage{
generate_model_from_modulenet(modulenet_name, platform, treeseed = NULL,
  decay = 1.4, ngenes_per_module_sampler = function(n_genes, n_modules)
  sample(1:10, n_modules, replace = TRUE), edge_retainment = function(n)
  max(c(round(n/2), 1)), ntargets_sampler = function(n_genes, n_regulators) { 
     sample(20:100, 1) }, main_targets_ratio = 0.05, target_adder_name,
  realnet_name = "regulatorycircuits", damping = 0.05, samplers,
  verbose = TRUE)
}
\arguments{
\item{modulenet_name}{A modulenet_name as defined in data/modulenetworks or `tree` to generate a random tree}

\item{platform}{The platform}

\item{treeseed}{A seed to set}

\item{decay}{The decay rate}

\item{ngenes_per_module_sampler}{Function for number of genes per module, given n_genes and n_modules}

\item{edge_retainment}{Function for sampling the number of edges retained between tfs of module nodes}

\item{ntargets_sampler}{Function for number of targets per regulator, given n_genes and n_regulators}

\item{main_targets_ratio}{Ratio of targets versus main tfs}

\item{target_adder_name}{The method for adding targets. Only "realnet" is currently supported.}

\item{realnet_name}{Name of the real network.}

\item{damping}{A daping factor for personalized pagerank}

\item{samplers}{The samplers for the kinetics parameters}

\item{verbose}{Whether or not to produce textual output}
}
\description{
Default parameters are defined in individual functions and are inherited in `zzz_param_inheritance.R`
}
