% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_kinetics.R
\name{generate_system}
\alias{generate_system}
\title{Generate the system}
\usage{
generate_system(net, geneinfo, cells,
  samplers = get_default_kinetics_samplers())
}
\arguments{
\item{net}{Regulatory network dataframe}

\item{geneinfo}{Geneinfo dataframe}

\item{cells}{Cells dataframe}

\item{samplers}{The samplers for the kinetics parameters}
}
\description{
Generate the system
}
