% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timings.R
\name{get_timings}
\alias{get_timings}
\title{Return the timings of each of the dyngen steps}
\usage{
get_timings(model)
}
\arguments{
\item{model}{A dyngen object}
}
\value{
A data frame with columns \code{"group"}, \code{"task"}, \code{"time_elapsed"}.
}
\description{
Return the timings of each of the dyngen steps
}
\examples{
data("example_model")
timings <- get_timings(example_model)
}
