% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_initialisation.R
\name{initialise_model}
\alias{initialise_model}
\title{Initial settings for simulating a dyngen dataset}
\usage{
initialise_model(num_cells = round(10^runif(1, 2, 4)),
  num_tfs = round(10^runif(1, 1, log10(200))),
  num_targets = round(10^runif(1, 1, 4)), num_hks = round(10^runif(1,
  1, 3)), distance_metric = c("pearson", "spearman", "cosine",
  "euclidean", "manhattan"), backbone,
  tf_network_params = tf_network_random(),
  feature_network_params = feature_network_default(),
  kinetics_params = kinetics_default(),
  simulation_params = simulation_default(),
  gold_standard_params = gold_standard_default(),
  experiment_params = experiment_snapshot(), verbose = TRUE,
  download_cache_dir = NULL, num_cores = 8)
}
\arguments{
\item{num_cells}{The number of cells to sample.}

\item{num_tfs}{The number of transcription factors to generate.}
}
\description{
Initial settings for simulating a dyngen dataset
}
