% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_initialisation.R
\name{initialise_model}
\alias{initialise_model}
\title{Initial settings for simulating a dyngen dataset}
\usage{
initialise_model(num_cells, num_tfs, num_targets, num_hks,
  dist_metric = c("spearman", "pearson", "cosine", "euclidean",
  "manhattan"), modulenet = modulenet_linear(),
  tf_network_params = tf_network_random(),
  feature_network_params = feature_network_realnet(),
  kinetics_params = kinetics_default(),
  simulation_params = simulation_default(),
  gold_standard_params = gold_standard_default(),
  experiment_params = experiment_snapshot(), verbose = FALSE,
  num_cores = 1, download_cache_dir = NULL)
}
\arguments{
\item{num_cells}{The number of cells to sample.}

\item{num_tfs}{The number of transcription factors to generate.}
}
\description{
Initial settings for simulating a dyngen dataset
}
