% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_loading_saving.R
\name{load_dataset}
\alias{load_dataset}
\title{Load a dataset}
\usage{
load_dataset(datasetid, contents = list(simulation = FALSE, goldstandard =
  TRUE, model = TRUE, experiment = TRUE))
}
\description{
Load a dataset
}
