% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_network_generation.R
\name{load_modulenet}
\alias{load_modulenet}
\title{load module net}
\usage{
load_modulenet(modulenetname)
}
\arguments{
\item{modulenetname}{A modulenetname as defined in data/modulenetworks or `tree` to generate a random tree}
}
\description{
load module net
}
