% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_network_generation.R
\name{load_modulenet}
\alias{load_modulenet}
\title{Load module network from extra data}
\usage{
load_modulenet(modulenet_name)
}
\arguments{
\item{modulenet_name}{A modulenet_name as defined in data/modulenetworks or `tree` to generate a random tree}
}
\description{
Load module network from extra data
}
