% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_estimate_platform.R
\name{get_platform_from_counts}
\alias{get_platform_from_counts}
\alias{get_simple_platform}
\title{Estimate a platform from a dataset}
\usage{
get_platform_from_counts(counts, grouping, subsample = 500)

get_simple_platform(n_cells = 100L, n_features = 100L,
  trajectory_dependent_features = 0.1, dropout_rate = 0.01,
  dropout_shape = 1)
}
\arguments{
\item{counts}{The counts with cells in rows and genes in columns.}

\item{grouping}{A named vector representing a grouping of the cells.}

\item{subsample}{The number of cells to subsample.}

\item{n_cells}{The number of cells}

\item{n_features}{The number of features}

\item{trajectory_dependent_features}{The percentage of features that are being driven by the trajectory (or vice versa)}

\item{dropout_rate}{The mean rate of dropouts}

\item{dropout_shape}{The shape of dropouts}
}
\description{
Altenatively, \code{\link[=get_simple_platform]{get_simple_platform()}} returns a simple platform parameter configuration.
}
