% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gold_expression}
\alias{plot_gold_expression}
\title{Visualise the expression of the gold standard over simulation time}
\usage{
plot_gold_expression(model, what = c("w", "x", "y"))
}
\arguments{
\item{model}{A dyngen intermediary model for which the simulations have been run with \code{\link[=generate_gold_standard]{generate_gold_standard()}}.}

\item{what}{Which molecule types to visualise.}
}
\description{
Visualise the expression of the gold standard over simulation time
}
