% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_goldstandard}
\alias{plot_goldstandard}
\alias{plot_goldstandard_edges}
\alias{plot_goldstandard_burn}
\alias{plot_goldstandard_network}
\alias{plot_goldstandard_heatmap}
\alias{plot_goldstandard_references}
\title{Gold standard plotting}
\usage{
plot_goldstandard(simulation, gs)

plot_goldstandard_edges(simulation, gs,
  spaces = dimred_goldstandard(simulation, gs))

plot_goldstandard_burn(simulation, gs,
  spaces = dimred_goldstandard(simulation, gs))

plot_goldstandard_network(simulation, gs,
  spaces = dimred_goldstandard(simulation, gs))

plot_goldstandard_heatmap(simulation, gs)

plot_goldstandard_references(gs)
}
\arguments{
\item{simulation}{The simulation}

\item{gs}{The gold standard}

\item{spaces}{Spaces dataframe}
}
\description{
Gold standard plotting
}
