% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_model}
\alias{plot_model}
\alias{plot_modulenet}
\alias{plot_net}
\alias{plot_net_overlaps}
\title{Model plotting}
\usage{
plot_model(model)

plot_modulenet(model)

plot_net(model, colorby = c("module", "main"), main_only = TRUE,
  label = FALSE)

plot_net_overlaps(model)
}
\arguments{
\item{model}{The model}

\item{colorby}{By what to color}

\item{main_only}{Whether to only draw the main network}

\item{label}{Whether to label genes}
}
\description{
Model plotting
}
