% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_goldstandard.R
\name{preprocess_simulation_for_gs}
\alias{preprocess_simulation_for_gs}
\title{Preprocessing simulation data for gold standard generation
Averaging over modules & smoothing expression}
\usage{
preprocess_simulation_for_gs(simulation, model, smooth_window = 50)
}
\arguments{
\item{simulation}{The simulation}

\item{model}{The model}

\item{smooth_window}{The window (steps) to smooth over}
}
\description{
Preprocessing simulation data for gold standard generation
Averaging over modules & smoothing expression
}
