% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy-2-cells.R
\name{random_progressions_tented}
\alias{random_progressions_tented}
\title{Assign random progressions with tenting}
\usage{
random_progressions_tented(milestone_network, ncells = 100)
}
\arguments{
\item{milestone_network}{The milestone network}

\item{ncells}{The number of cells to generate}
}
\description{
Assign random progressions with tenting
}
