% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_kinetics.R
\name{randomize_gene_kinetics}
\alias{randomize_gene_kinetics}
\title{Randomize kinetics of genes}
\usage{
randomize_gene_kinetics(geneinfo, net)
}
\arguments{
\item{geneinfo}{Geneinfo dataframe}

\item{net}{Regulatory network dataframe}
}
\description{
Randomize kinetics of genes
}
