% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_experiment.R
\name{run_experiment}
\alias{run_experiment}
\title{Run an experiment from a simulation}
\usage{
run_experiment(simulation, gs, sampler, platform)
}
\arguments{
\item{simulation}{The simulation}

\item{gs}{The gold standard}

\item{sampler}{Function telling how the cells should be sampled}

\item{platform}{Platform}
}
\description{
Run an experiment from a simulation
}
