% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtime_from_backbone.R
\name{simtime_from_backbone}
\alias{simtime_from_backbone}
\title{Determine simulation time from backbone}
\usage{
simtime_from_backbone(backbone, burn = FALSE)
}
\arguments{
\item{backbone}{A valid dyngen backbone object}

\item{burn}{Whether or not to compute the simtime for only the burn phase}
}
\description{
Determine simulation time from backbone
}
