% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_transform.R
\name{wrap_dataset}
\alias{wrap_dataset}
\title{Wrap a simulation into a dynwrap object}
\usage{
wrap_dataset(
  model,
  store_dimred = !is.null(model$simulations$dimred),
  store_cellwise_grn = !is.null(model$experiment$cellwise_grn),
  store_rna_velocity = !is.null(model$experiment$rna_velocity)
)
}
\arguments{
\item{model}{A dyngen output model for which the experiment has been emulated with \code{\link[=generate_experiment]{generate_experiment()}}.}

\item{store_dimred}{Whether or not to store the dimensionality reduction constructed on the true counts.}

\item{store_cellwise_grn}{Whether or not to also store cellwise GRN information.}

\item{store_rna_velocity}{WHether or not to store the log propensity ratios.}
}
\value{
A dynwrap object.
}
\description{
The output of this object can be used with \pkg{dyno}.
}
\examples{
model <- initialise_model(
  backbone = backbone_bifurcating()
)
\dontshow{
# actually use a smaller example 
# to reduce execution time during
# testing of the examples
model <- initialise_model(
  backbone = model$backbone,
  num_cells = 5,
  num_targets = 0,
  num_hks = 0,
  gold_standard_params = gold_standard_default(census_interval = 1, tau = 0.1),
  simulation_params = simulation_default(
    burn_time = 10,
    total_time = 10,
    census_interval = 1,
    ssa_algorithm = ssa_etl(tau = 0.1),
    experiment_params = simulation_type_wild_type(num_simulations = 1)
  )
)
}
\donttest{
model <- model \%>\%
  generate_tf_network() \%>\%
  generate_feature_network() \%>\%
  generate_kinetics() \%>\%
  generate_gold_standard() \%>\%
  generate_cells() \%>\%
  generate_experiment()
  
dataset <- wrap_dataset(model)

# dynplot::plot_dimred(dataset)
}
}
