% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_dynwrap.R
\name{wrap_dataset}
\alias{wrap_dataset}
\title{Wrap a simulation into a dynwrap object}
\usage{
wrap_dataset(model, store_grn = FALSE, store_dimred = FALSE)
}
\arguments{
\item{model}{A dyngen output model for which the experiment has been emulated with \code{\link[=generate_experiment]{generate_experiment()}}.}

\item{store_grn}{Whether or not to also store (cellwise) GRN information.}

\item{store_dimred}{Whether or not to store the dimensionality reduction constructed on the true counts.}
}
\description{
The output of this object can be used with \pkg{dyno}.
}
