% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_wraptasks.R
\name{wrap_task}
\alias{wrap_task}
\title{Wrap tasks}
\usage{
wrap_task(params, model, simulation, gs, experiment, normalization)
}
\arguments{
\item{params}{Parameters of the dataset}

\item{model}{Model}

\item{simulation}{Simulation}

\item{gs}{Gold standard}

\item{experiment}{Experiment}

\item{normalization}{Normalization}
}
\description{
Wrap tasks
}
